Make Edit possible
implemented Environment and State Object with Wrappers to make edit possible and make all views change.
