+++ iOS/Financer/Financer/Views/LegalPerson/EditLegalPerson.swift
@@ -7,9 +7,43 @@
 
 import SwiftUI
 
-struct EditLegalPerson: View {
+/// The View to edit a Legal Person
+internal struct EditLegalPerson: View {
+    
+    /// The View Context to interact with Core Data
+    @Environment(\.managedObjectContext) private var viewContext
+    
+    /// The LegalPersonWrapper in the Environment
+    @EnvironmentObject private var legalPersonWrapper : LegalPersonWrapper
+    
+    /// Whether ther Error when saving is displayed or not.
+    @State private var errSavingPresented : Bool = false
+    
     var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+        LegalPersonEditor(action: editLegalPerson, legalPerson: legalPersonWrapper.legalPerson!)
+            .navigationTitle("Edit \(legalPersonWrapper.legalPerson!.name!)")
+            .alert(
+                "Error",
+                isPresented: $errSavingPresented
+            ) {
+                
+            } message: {
+                Text(
+                    "Error saving Data.\nPlease try again\n\nIf this Error occurs again, please contact the support."
+                )
+            }
+    }
+    
+    /// The action that is passed to the Legal Person
+    /// Editor and edits a Legal Person
+    private func editLegalPerson(legalPerson : LegalPerson) -> Void {
+        viewContext.delete(legalPersonWrapper.legalPerson!)
+        legalPersonWrapper.legalPerson = legalPerson
+        do {
+            try viewContext.save()
+        } catch _ {
+            errSavingPresented.toggle()
+        }
     }
 }
