+++ iOS/Financer/Financer/Views/LegalPerson/LegalPersonDetails.swift
@@ -16,18 +16,20 @@ internal struct LegalPersonDetails: View {
     @FetchRequest private var finances : FetchedResults<Finance>
     
     /// The Legal Parson beeing shown right here
-    @State private var legalPerson : LegalPerson?
+    @EnvironmentObject private var legalPersonWrapper : LegalPersonWrapper
     
-    /// The Initializer with the Legal Person this Screen should be
-    /// generated for.
-    internal init(legalPerson: LegalPerson) {
-        self._legalPerson = State(initialValue:  legalPerson)
+    /// The non optional Legal Person to show in this View
+    private let legalPerson : LegalPerson
+    
+    /// The Standard Initializer for this View
+    internal init() {
+        legalPerson = legalPersonWrapper.legalPerson!
         _finances = FetchRequest(
             entity: Finance.entity(),
             sortDescriptors: [
                 NSSortDescriptor(keyPath: \Finance.date, ascending: true)
             ],
-            predicate: NSPredicate(format: "legalPerson == %@", legalPerson)
+            predicate: NSPredicate(format: "legalPerson == %@", legalPersonWrapper.legalPerson!)
         )
     }
     
@@ -35,15 +37,15 @@ internal struct LegalPersonDetails: View {
         NavigationStack {
             List {
                 Section {
-                    DefaultListTile(name: "Name", data: legalPerson!.name!)
-                    DefaultListTile(name: "Type", data: legalPerson!.typeAsString())
+                    DefaultListTile(name: "Name", data: legalPerson.name!)
+                    DefaultListTile(name: "Type", data: legalPerson.typeAsString())
                 } header: {
                     Text("General Values")
                 } footer: {
-                    Text("These are the general Values for this \(legalPerson!.name!)")
+                    Text("These are the general Values for this \(legalPerson.name!)")
                 }
                 Section {
-                    DefaultListTile(name: "Phone", data: legalPerson!.phone!)
+                    DefaultListTile(name: "Phone", data: legalPerson.phone!)
                     Text(notes)
                         .lineLimit(5...10)
                         .foregroundColor(.gray)
@@ -65,14 +67,15 @@ internal struct LegalPersonDetails: View {
                     Text("Represents all relations this Finance has.")
                 }
             }
-            .navigationTitle("\(legalPerson!.name!) Details")
+            .navigationTitle("\(legalPerson.name!) Details")
             .navigationBarTitleDisplayMode(.automatic)
             .toolbarRole(.navigationStack)
             .toolbar(.automatic, for: .navigationBar)
             .toolbar {
                 ToolbarItem(placement: .primaryAction) {
                     NavigationLink {
-                        AddLegalPerson(legalPerson: $legalPerson)
+                        EditLegalPerson()
+                            .environmentObject(legalPersonWrapper)
                     } label: {
                         Image(systemName: "pencil")
                     }
@@ -107,10 +110,10 @@ internal struct LegalPersonDetails: View {
     /// this returns an information String
     /// stating exactly that.
     private var notes : String {
-        if legalPerson!.notes!.isEmpty {
+        if legalPerson.notes!.isEmpty {
             return "No Notes"
         } else {
-            return legalPerson!.notes!
+            return legalPerson.notes!
         }
     }
 }
@@ -118,9 +121,10 @@ internal struct LegalPersonDetails: View {
 internal struct LegalPersonDetails_Previews: PreviewProvider {
     
     /// The preview legal Person
-    private static var person : LegalPerson = LegalPerson.anonymous
+    private static var personWrapper : LegalPersonWrapper = LegalPersonWrapper(legalPerson: LegalPerson.anonymous)
     
     static var previews: some View {
-        LegalPersonDetails(legalPerson: person)
+        LegalPersonDetails()
+            .environmentObject(personWrapper)
     }
 }
